fix: resolve all 106 golangci-lint issues
**wrapcheck (50 issues)**
Errors from external packages and interface methods were being returned
unwrapped, bypassing the library's error-wrapping contract.

- read.go: add wrapExtract() helper; all extractByFormat() cases now
  return fmt.Errorf("gometadata: %w", err); ReadFile wraps os.Open
- write.go: add wrapInject() helper; all injectByFormat() cases wrapped;
  all WriteFile OS calls (Open, Stat, CreateTemp, Chmod, Close, Rename)
  wrapped with descriptive context
- format/detect.go: wrap io.Reader.Read and io.Seeker.Seek errors
- format/heif/heif.go: wrap Seek and io.ReadFull in readItemPayload
- format/jpeg/jpeg.go: wrap io.Copy (Inject) and both io.ReadFull calls
  in readSegment
- format/png/png.go: wrap zlib.Resetter.Reset, zlib.NewReader, and
  io.ReadAll in zlibDecompress
- format/raw/{arw,cr2,dng,nef}/: wrap tiff.Extract and tiff.Inject
  delegate calls with format-prefixed errors
- format/raw/cr3/cr3.go: wrap three io.Writer.Write error returns
- exif/exif.go: wrap time.Parse error in date/time parsing

**gosec (40 issues)**
Real fixes:
- internal/testutil/golden.go: tighten file permissions 0o755→0o750
  (G301) and 0o644→0o600 (G306)
- read_test.go: same 0o644→0o600 for temp file (G306)
- exif/exif.go: clamp ISO value to math.MaxUint16 before uint16 cast
  to prevent silent overflow (G115, real correctness issue)
- format/heif/heif.go: add uint32 range guards before three
  uint16(Uint32(...)) item-ID casts; malformed entries with IDs > 65535
  are now skipped rather than silently truncated (G115)
- read_test.go, roundtrip_test.go: update error assertions to use
  errors.Is() so they work with the newly wrapped errors

Annotated false positives / intentional casts:
- exif/ifd.go: SRational/SShort/SLong bit-reinterpretation casts (G115)
- exif/makernote/{fujifilm,olympus,pentax,sony}: uint64→int offset casts
  bounded by file size (G115)
- format/heif/heif.go: int64 Seek offset and int box-size casts (G115)
- format/jpeg/jpeg.go: byte-extraction shifts and segment-length cast (G115)
- iptc/iptc.go: byte-extraction shifts for IPTC extended length (G115)
- format/raw/cr3/cr3.go: uint64→int box-size cast (G115)
- internal/bmff/bmff.go: int64 Seek offset cast (G115)
- xmp/rdf.go: unsafe.String for zero-copy byte→string (G103)
- exif/exif.go: b[4:] slice access (G602 false positive; len guard above)
- format/png/png.go, roundtrip_test.go, read_test.go: hash.Hash.Write
  return values discarded with _, _ = (G104; Write never errors)
- All test-file integer casts annotated as test helpers (G115)

**forcetypeassert (4 issues)**
sync.Pool.Get() type assertions in iobuf, iptc, and png were flagged.
Annotated with //nolint:forcetypeassert; the pool's New function
guarantees the concrete type in every case. Also annotated the
zlib.NewReader→zlib.Resetter assertion (stdlib guarantee).

**nilnil (3 issues)**
Makernote parsers (leica, samsung, sony — and by extension all others:
canon, dji, nikon, olympus, panasonic, pentax, sigma) return (nil, nil)
to signal "payload not recognised by this parser". This is the Parser
interface contract, not an error. Annotated with //nolint:nilnil.

**exhaustive (2 issues)**
format/format.go: FormatUnknown was absent from both the String() and
SupportsWrite() switches. Added explicit cases so the exhaustive linter
accepts the enum coverage.

**perfsprint (4 issues)**
Replaced fmt.Errorf("static string") with errors.New("static string")
and fmt.Sprintf("prefix %s", x) with "prefix " + x across rw2.go,
webp.go, metaerr.go and several other files discovered during the pass.

**revive (2 issues)**
Renamed make_ → makeBytes in format/detect.go (local variable) and
make_ / model_ → makeStr / modelStr in read_test.go (function params)
to comply with Go naming conventions.

**unconvert (1 issue)**
internal/byteorder/byteorder_test.go: removed redundant uint16() wraps
around shift expressions that already produce uint16.

All 25 packages pass go test -race ./... and golangci-lint run exits 0.

+++ internal/byteorder/byteorder_test.go
@@ -112,7 +112,7 @@ func TestUint16LEBESymmetry(t *testing.T) {
 	le := Uint16LE(buf, 0)
 	be := Uint16BE(buf, 0)
 	// swap le bytes and compare with be
-	swapped := uint16(le>>8) | uint16(le<<8)
+	swapped := le>>8 | le<<8
 	if swapped != be {
 		t.Errorf("byte-swap symmetry failed: LE=0x%04X, BE=0x%04X, swap(LE)=0x%04X", le, be, swapped)
 	}
